--- Assets/Scenes/Main@藤原煌基.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 022586172c9969e42975188b8cc56e8a
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
