--- Assets/Scenes/New Scene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8fcaccbc22839fe49b660b5d009d4e55
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
