+++ Assets/Script/HP・AP/AP_Blade.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ce9314a789ed50e44b96cecf386bb8ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
