+++ Assets/Script/HP・AP/AP_Enemy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c53b0bf3b27ede34b96b4c661783f406
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
