--- Assets/Script/HP・AP/AP_Enmy.cs
@@ -1,51 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class AP_Enmy : MonoBehaviour
-{
-    public float damageAmount = 10f;
-    // Start is called before the first frame update
-    void Start()
-    {
-
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-
-
-    private void OnTriggerEnter(Collider collision)
-    {
-        HP health = collision.gameObject.GetComponent<HP>();
-
-        if (health != null)
-        {
-            // ƒ_ƒ[ƒW‚ğ—^‚¦‚é
-            health.TakeDamage(damageAmount);
-        }
-
-        if (collision.gameObject.CompareTag("Earth"))//collision‚É“ü‚Á‚Ä‚¢‚éTag‚ªEarth‚Ì
-        {
-            Destroy(this.gameObject);
-        }
-
-        if (collision.gameObject.CompareTag("Enemy"))//collision‚É“ü‚Á‚Ä‚¢‚éTag‚ªEnemy‚Ì
-        {
-            Destroy(this.gameObject);
-        }
-
-        if (collision.gameObject.CompareTag("Bullet_P"))//collision‚É“ü‚Á‚Ä‚¢‚éTag‚ªEarth‚Ì
-        {
-            Destroy(this.gameObject);
-        }
-
-        if (collision.gameObject.CompareTag("Laser"))//collision‚É“ü‚Á‚Ä‚¢‚éTag‚ªEarth‚Ì
-        {
-            Destroy(this.gameObject);
-        }
-    }
-}
