+++ Assets/Script/HP・AP/AP_Laser.cs
@@ -20,7 +20,7 @@ public class AP_Laser : MonoBehaviour
 
     private void OnTriggerEnter(Collider collision)
     {
-        HP health = collision.gameObject.GetComponent<HP>();
+        HP_Earth health = collision.gameObject.GetComponent<HP_Earth>();
 
         if (health != null)
         {
@@ -29,11 +29,6 @@ public class AP_Laser : MonoBehaviour
         }
 
         if (collision.gameObject.CompareTag("Earth"))//collision‚É“ü‚Á‚Ä‚¢‚éTag‚ªEarth‚ÌŽž
-        {
-            Destroy(this.gameObject);
-        }
-
-        if (collision.gameObject.CompareTag("Enemy"))//collision‚É“ü‚Á‚Ä‚¢‚éTag‚ªEnemy‚ÌŽž
         {
             Destroy(this.gameObject);
         }
