--- Assets/missile.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: fb8765ca41998dc458bc18a01d23c812
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
