+++ Assets/use prefab/Bomb.prefab
@@ -134,7 +134,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 67d7eb8931b25d443a53d8b24cab4d52, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  damageAmount: 30
+  damageAmount: 20
 --- !u!54 &-607135113825295115
 Rigidbody:
   m_ObjectHideFlags: 0
