+++ Assets/use prefab/Laser.prefab
@@ -147,7 +147,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 096f5bd77e59d3e41956ad92b6af36d6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  damageAmount: 50
+  damageAmount: 25
 --- !u!114 &-1527486519127396032
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -160,5 +160,5 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f1d0df61e1eeeda4aae2928d4cd136e4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  SpeedX: 100
+  SpeedX: 85
   LifeTime: 2
