+++ Assets/素材.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 837e0b9ffa7269440a0d245a94e12740
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
